+++ changelog.md
@@ -1,5 +1,11 @@
 # Rockup Go Standard Library Changelog
 
+## v1.1.1
+12/05/2024
+- fix concurrent map read/write bug in static.go
+- add web.FileCSV response type to package web
+
+
 ## v1.1.0 - go1.22
 09/02/2024
 - added package v(alidate)